# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentación Taller Digicom By: Richard Ortiz",
    "version": "1.0.0",
    "description": "Documentación de la API para la gestión de talleres"
  },
  "servers": [
    {
      "url": "https://api.taller.digicom.com.gt/api/v1",
      "description": "Production API over HTTPS"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development API"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
    "/usuarios": {
      "get": {
        "summary": "Obtiene una lista de usuarios",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer"
            },
            "default": 20
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "idUsuarios": {
                        "type": "integer"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ningún usuario localizado."
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      },
      "post": {
        "summary": "Crea un nuevo usuario",
        "tags": [
          "Usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "senha": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario añadido con éxito"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
...
  },
...
}